# perf: use `params ReadOnlySpan` for `List.Add`